Pose Detection Model
This will detect the body parts in an uploaded image so it can be segmented for wedding dress fitting. In the beginning of the code are a list of body parts and their points they correspond to. 

You will probably have to segement the body part numbers where the lines are by looking at the corresponding lines and placement of data points (red dots). Look in the beginning of the code and start there.